rafeactor: better code organization
