+++ src/lib/route-processor.ts
@@ -1,29 +1,25 @@
-import * as t from "@babel/types";
 import fs from "fs";
 import path from "path";
-import traverseModule from "@babel/traverse";
-
-// Handle both ES modules and CommonJS
-const traverse = (traverseModule as any).default || traverseModule;
 
 import { SchemaProcessor } from "./schema-processor.js";
 import {
   capitalize,
-  extractJSDocComments,
-  parseTypeScriptFile,
   extractPathParameters,
   getOperationId,
 } from "./utils.js";
 import { DataTypes, OpenApiConfig, RouteDefinition } from "../types.js";
 import { logger } from "./logger.js";
+import { RouterStrategy } from "./router-strategy.js";
+import { AppRouterStrategy } from "./app-router-strategy.js";
+import { PagesRouterStrategy } from "./pages-router-strategy.js";
 
-const HTTP_METHODS = ["GET", "POST", "PUT", "PATCH", "DELETE"];
 const MUTATION_HTTP_METHODS = ["PATCH", "POST", "PUT"];
 
 export class RouteProcessor {
   private swaggerPaths: Record<string, any> = {};
   private schemaProcessor: SchemaProcessor;
   private config: OpenApiConfig;
+  private strategy: RouterStrategy;
 
   private directoryCache: Record<string, string[]> = {};
   private statCache: Record<string, fs.Stats> = {};
@@ -36,6 +32,9 @@ export class RouteProcessor {
       config.schemaType,
       config.schemaFiles
     );
+    this.strategy = config.routerType === "pages"
+      ? new PagesRouterStrategy(config)
+      : new AppRouterStrategy(config);
   }
 
   private buildResponsesFromConfig(
@@ -194,10 +193,6 @@ export class RouteProcessor {
     return this.schemaProcessor;
   }
 
-  private isRoute(varName: string): boolean {
-    return HTTP_METHODS.includes(varName);
-  }
-
   /**
    * Check if a route should be ignored based on config patterns or @ignore tag
    */
@@ -221,380 +216,31 @@ export class RouteProcessor {
     });
   }
 
-  private processFile(filePath: string): void {
-    // Check if the file has already been processed
-    if (this.processFileTracker[filePath]) return;
-
-    const content = fs.readFileSync(filePath, "utf-8");
-    const ast = parseTypeScriptFile(content);
-
-    traverse(ast, {
-      ExportNamedDeclaration: (path) => {
-        const declaration = path.node.declaration;
-
-        if (
-          t.isFunctionDeclaration(declaration) &&
-          t.isIdentifier(declaration.id)
-        ) {
-          const dataTypes = extractJSDocComments(path);
-          if (this.isRoute(declaration.id.name)) {
-            const routePath = this.getRoutePath(filePath);
-
-            // Skip if route should be ignored
-            if (this.shouldIgnoreRoute(routePath, dataTypes)) {
-              logger.debug(`Ignoring route: ${routePath}`);
-              return;
-            }
-
-            // Don't bother adding routes for processing if only including OpenAPI routes and the route is not OpenAPI
-            if (
-              !this.config.includeOpenApiRoutes ||
-              (this.config.includeOpenApiRoutes && dataTypes.isOpenApi)
-            ) {
-              // Check for URL parameters in the route path
-              const pathParams = extractPathParameters(routePath);
-
-              // If we have path parameters but no pathParamsType defined, we should log a warning
-              if (pathParams.length > 0 && !dataTypes.pathParamsType) {
-                logger.debug(
-                  `Route ${routePath} contains path parameters ${pathParams.join(
-                    ", "
-                  )} but no @pathParams type is defined.`
-                );
-              }
-
-              this.addRouteToPaths(declaration.id.name, filePath, dataTypes);
-            }
-          }
-        }
-
-        if (t.isVariableDeclaration(declaration)) {
-          declaration.declarations.forEach((decl) => {
-            if (t.isVariableDeclarator(decl) && t.isIdentifier(decl.id)) {
-              if (this.isRoute(decl.id.name)) {
-                const dataTypes = extractJSDocComments(path);
-                const routePath = this.getRoutePath(filePath);
-
-                // Skip if route should be ignored
-                if (this.shouldIgnoreRoute(routePath, dataTypes)) {
-                  logger.debug(`Ignoring route: ${routePath}`);
-                  return;
-                }
-
-                // Don't bother adding routes for processing if only including OpenAPI routes and the route is not OpenAPI
-                if (
-                  !this.config.includeOpenApiRoutes ||
-                  (this.config.includeOpenApiRoutes && dataTypes.isOpenApi)
-                ) {
-                  const pathParams = extractPathParameters(routePath);
-
-                  if (pathParams.length > 0 && !dataTypes.pathParamsType) {
-                    logger.debug(
-                      `Route ${routePath} contains path parameters ${pathParams.join(
-                        ", "
-                      )} but no @pathParams type is defined.`
-                    );
-                  }
-
-                  this.addRouteToPaths(decl.id.name, filePath, dataTypes);
-                }
-              }
-            }
-          });
-        }
-      },
-    });
-
-    this.processFileTracker[filePath] = true;
-  }
-
   /**
-   * Process Pages Router API files
-   * Pages Router uses export default function handler with @method tags
+   * Register a discovered route after filtering
    */
-  private processPagesRouterFile(filePath: string): void {
-    // Check if the file has already been processed
-    if (this.processFileTracker[filePath]) return;
+  private registerRoute(method: string, filePath: string, dataTypes: DataTypes): void {
+    const routePath = this.strategy.getRoutePath(filePath);
 
-    const content = fs.readFileSync(filePath, "utf-8");
-    const ast = parseTypeScriptFile(content);
-
-    // Collect all JSDoc comments with @method tags
-    const methodComments: { method: string; dataTypes: DataTypes }[] = [];
-
-    traverse(ast, {
-      ExportDefaultDeclaration: (nodePath) => {
-        // Get all leading comments for the export default
-        const allComments = ast.comments || [];
-        const exportStart = nodePath.node.start || 0;
-
-        // Find all block comments before this export that contain @method
-        allComments.forEach((comment) => {
-          if (
-            comment.type === "CommentBlock" &&
-            (comment.end || 0) < exportStart
-          ) {
-            const commentValue = comment.value;
-            if (commentValue.includes("@method")) {
-              // Parse this comment block
-              const dataTypes = this.extractJSDocFromComment(commentValue);
-              if (dataTypes.method && HTTP_METHODS.includes(dataTypes.method)) {
-                methodComments.push({
-                  method: dataTypes.method,
-                  dataTypes,
-                });
-              }
-            }
-          }
-        });
-
-        // Process each method found
-        methodComments.forEach(({ method, dataTypes }) => {
-          const routePath = this.getPagesRoutePath(filePath);
-
-          // Skip if route should be ignored
-          if (this.shouldIgnoreRoute(routePath, dataTypes)) {
-            logger.debug(`Ignoring route: ${routePath} [${method}]`);
-            return;
-          }
-
-          // Don't bother adding routes for processing if only including OpenAPI routes and the route is not OpenAPI
-          if (
-            !this.config.includeOpenApiRoutes ||
-            (this.config.includeOpenApiRoutes && dataTypes.isOpenApi)
-          ) {
-            const pathParams = extractPathParameters(routePath);
-
-            if (pathParams.length > 0 && !dataTypes.pathParamsType) {
-              logger.debug(
-                `Route ${routePath} contains path parameters ${pathParams.join(
-                  ", "
-                )} but no @pathParams type is defined.`
-              );
-            }
-
-            this.addRouteToPaths(method, filePath, dataTypes);
-          }
-        });
-      },
-    });
-
-    this.processFileTracker[filePath] = true;
-  }
-
-  /**
-   * Extract JSDoc data from a comment string (for Pages Router)
-   */
-  private extractJSDocFromComment(commentValue: string): DataTypes {
-    const cleanedComment = commentValue.replace(/\*\s*/g, "").trim();
-
-    let tag = "";
-    let summary = "";
-    let description = "";
-    let paramsType = "";
-    let pathParamsType = "";
-    let bodyType = "";
-    let auth = "";
-    let isOpenApi = cleanedComment.includes("@openapi");
-    let isIgnored = cleanedComment.includes("@ignore");
-    let deprecated = cleanedComment.includes("@deprecated");
-    let bodyDescription = "";
-    let contentType = "";
-    let responseType = "";
-    let responseDescription = "";
-    let responseSet = "";
-    let addResponses = "";
-    let successCode = "";
-    let operationId = "";
-    let method = "";
-
-    // Extract @method
-    const methodMatch = cleanedComment.match(/@method\s+(\S+)/);
-    if (methodMatch) {
-      method = methodMatch[1].trim().toUpperCase();
+    if (this.shouldIgnoreRoute(routePath, dataTypes)) {
+      logger.debug(`Ignoring route: ${routePath}`);
+      return;
     }
 
-    // Extract summary (first line that doesn't start with @)
-    const firstLine = cleanedComment.split("\n")[0];
-    if (!firstLine.trim().startsWith("@")) {
-      summary = firstLine.trim();
+    if (this.config.includeOpenApiRoutes && !dataTypes.isOpenApi) {
+      return;
     }
 
-    // Extract @description
-    const descMatch = cleanedComment.match(/@description\s+(.*)/);
-    if (descMatch) {
-      description = descMatch[1].trim();
-    }
-
-    // Extract @tag
-    const tagMatch = cleanedComment.match(/@tag\s+(.*)/);
-    if (tagMatch) {
-      tag = tagMatch[1].trim();
-    }
-
-    // Extract @params or @queryParams
-    const paramsMatch =
-      cleanedComment.match(/@queryParams\s+([\w<>,\s\[\]]+)/) ||
-      cleanedComment.match(/@params\s+([\w<>,\s\[\]]+)/);
-    if (paramsMatch) {
-      paramsType = paramsMatch[1].trim();
-    }
-
-    // Extract @pathParams
-    const pathParamsMatch = cleanedComment.match(
-      /@pathParams\s+([\w<>,\s\[\]]+)/
-    );
-    if (pathParamsMatch) {
-      pathParamsType = pathParamsMatch[1].trim();
-    }
-
-    // Extract @body
-    const bodyMatch = cleanedComment.match(/@body\s+([\w<>,\s\[\]]+)/);
-    if (bodyMatch) {
-      bodyType = bodyMatch[1].trim();
-    }
-
-    // Extract @bodyDescription
-    const bodyDescMatch = cleanedComment.match(/@bodyDescription\s+(.*)/);
-    if (bodyDescMatch) {
-      bodyDescription = bodyDescMatch[1].trim();
-    }
-
-    // Extract @contentType
-    const contentTypeMatch = cleanedComment.match(/@contentType\s+(.*)/);
-    if (contentTypeMatch) {
-      contentType = contentTypeMatch[1].trim();
-    }
-
-    // Extract @response
-    const responseMatch = cleanedComment.match(
-      /@response\s+(?:(\d+):)?([^@\n\r]+)/
-    );
-    if (responseMatch) {
-      const [, code, type] = responseMatch;
-      const trimmedType = type?.trim();
-
-      if (!code && trimmedType && /^\d{3}$/.test(trimmedType)) {
-        successCode = trimmedType;
-        responseType = undefined;
-      } else {
-        successCode = code || "";
-        responseType = trimmedType;
-      }
-    }
-
-    // Extract @responseDescription
-    const respDescMatch = cleanedComment.match(/@responseDescription\s+(.*)/);
-    if (respDescMatch) {
-      responseDescription = respDescMatch[1].trim();
-    }
-
-    // Extract @responseSet
-    const respSetMatch = cleanedComment.match(/@responseSet\s+(.*)/);
-    if (respSetMatch) {
-      responseSet = respSetMatch[1].trim();
-    }
-
-    // Extract @add
-    const addMatch = cleanedComment.match(/@add\s+(.*)/);
-    if (addMatch) {
-      addResponses = addMatch[1].trim();
-    }
-
-    // Extract @operationId
-    const opIdMatch = cleanedComment.match(/@operationId\s+(\S+)/);
-    if (opIdMatch) {
-      operationId = opIdMatch[1].trim();
-    }
-
-    // Extract @auth
-    const authMatch = cleanedComment.match(/@auth\s+(.*)/);
-    if (authMatch) {
-      const authValue = authMatch[1].trim();
-      switch (authValue) {
-        case "bearer":
-          auth = "BearerAuth";
-          break;
-        case "basic":
-          auth = "BasicAuth";
-          break;
-        case "apikey":
-          auth = "ApiKeyAuth";
-          break;
-      }
-    }
-
-    return {
-      tag,
-      auth,
-      summary,
-      description,
-      paramsType,
-      pathParamsType,
-      bodyType,
-      isOpenApi,
-      isIgnored,
-      deprecated,
-      bodyDescription,
-      contentType,
-      responseType,
-      responseDescription,
-      responseSet,
-      addResponses,
-      successCode,
-      operationId,
-      method,
-    };
-  }
-
-  /**
-   * Get route path for Pages Router files
-   */
-  private getPagesRoutePath(filePath: string): string {
-    // Normalize path separators first
-    const normalizedPath = filePath.replaceAll("\\", "/");
-
-    // Normalize apiDir to ensure consistent format
-    const normalizedApiDir = this.config.apiDir
-      .replaceAll("\\", "/")
-      .replace(/^\.\//, "")
-      .replace(/\/$/, "");
-
-    // Find the apiDir position in the normalized path
-    const apiDirIndex = normalizedPath.indexOf(normalizedApiDir);
-
-    if (apiDirIndex === -1) {
-      throw new Error(
-        `Could not find apiDir "${this.config.apiDir}" in file path "${filePath}"`
+    const pathParams = extractPathParameters(routePath);
+    if (pathParams.length > 0 && !dataTypes.pathParamsType) {
+      logger.debug(
+        `Route ${routePath} contains path parameters ${pathParams.join(
+          ", "
+        )} but no @pathParams type is defined.`
       );
     }
 
-    // Extract the path after apiDir
-    let relativePath = normalizedPath.substring(
-      apiDirIndex + normalizedApiDir.length
-    );
-
-    // Remove the file extension (.ts or .tsx)
-    relativePath = relativePath.replace(/\.tsx?$/, "");
-
-    // Remove /index suffix (pages/api/users/index.ts -> /users)
-    relativePath = relativePath.replace(/\/index$/, "");
-
-    // Ensure the path starts with /
-    if (!relativePath.startsWith("/")) {
-      relativePath = "/" + relativePath;
-    }
-
-    // Remove trailing slash
-    relativePath = relativePath.replace(/\/$/, "");
-
-    // Handle catch-all routes ([...param]) before converting dynamic routes
-    relativePath = relativePath.replace(/\/\[\.\.\.(.*?)\]/g, "/{$1}");
-
-    // Convert Next.js dynamic route syntax to OpenAPI parameter syntax
-    relativePath = relativePath.replace(/\/\[([^\]]+)\]/g, "/{$1}");
-
-    return relativePath || "/";
+    this.addRouteToPaths(method, filePath, dataTypes);
   }
 
   public scanApiRoutes(dir: string): void {
@@ -606,8 +252,6 @@ export class RouteProcessor {
       this.directoryCache[dir] = files;
     }
 
-    const isPagesRouter = this.config.routerType === "pages";
-
     files.forEach((file) => {
       const filePath = path.join(dir, file);
       let stat = this.statCache[filePath];
@@ -618,18 +262,12 @@ export class RouteProcessor {
 
       if (stat.isDirectory()) {
         this.scanApiRoutes(filePath);
-        // @ts-ignore
-      } else if (file.endsWith(".ts") || file.endsWith(".tsx")) {
-        if (isPagesRouter) {
-          // Pages Router: process all .ts/.tsx files except _middleware, _app, _document
-          if (!file.startsWith("_")) {
-            this.processPagesRouterFile(filePath);
-          }
-        } else {
-          // App Router: only process route.ts/route.tsx files
-          if (file === "route.ts" || file === "route.tsx") {
-            this.processFile(filePath);
-          }
+      } else if (this.strategy.shouldProcessFile(file)) {
+        if (!this.processFileTracker[filePath]) {
+          this.strategy.processFile(filePath, (method, fp, dataTypes) => {
+            this.registerRoute(method, fp, dataTypes);
+          });
+          this.processFileTracker[filePath] = true;
         }
       }
     });
@@ -641,10 +279,7 @@ export class RouteProcessor {
     dataTypes: DataTypes
   ): void {
     const method = varName.toLowerCase();
-    const routePath =
-      this.config.routerType === "pages"
-        ? this.getPagesRoutePath(filePath)
-        : this.getRoutePath(filePath);
+    const routePath = this.strategy.getRoutePath(filePath);
     const rootPath = capitalize(routePath.split("/")[1]);
     const operationId = dataTypes.operationId || getOperationId(routePath, method);
     const {
@@ -769,54 +404,6 @@ export class RouteProcessor {
     }
 
     this.swaggerPaths[routePath][method] = definition;
-  }
-
-  private getRoutePath(filePath: string): string {
-    // Normalize path separators first
-    const normalizedPath = filePath.replaceAll("\\", "/");
-
-    // Normalize apiDir to ensure consistent format
-    const normalizedApiDir = this.config.apiDir
-      .replaceAll("\\", "/")
-      .replace(/^\.\//, "")
-      .replace(/\/$/, "");
-
-    // Find the apiDir position in the normalized path
-    const apiDirIndex = normalizedPath.indexOf(normalizedApiDir);
-
-    if (apiDirIndex === -1) {
-      throw new Error(
-        `Could not find apiDir "${this.config.apiDir}" in file path "${filePath}"`
-      );
-    }
-
-    // Extract the path after apiDir
-    let relativePath = normalizedPath.substring(
-      apiDirIndex + normalizedApiDir.length
-    );
-
-    // Remove the /route.ts or /route.tsx suffix
-    relativePath = relativePath.replace(/\/route\.tsx?$/, "");
-
-    // Ensure the path starts with /
-    if (!relativePath.startsWith("/")) {
-      relativePath = "/" + relativePath;
-    }
-
-    // Remove trailing slash
-    relativePath = relativePath.replace(/\/$/, "");
-
-    // Remove Next.js route groups (folders in parentheses like (authenticated), (marketing))
-    relativePath = relativePath.replace(/\/\([^)]+\)/g, "");
-
-    // Handle catch-all routes ([...param]) before converting dynamic routes
-    // This must come first because [...param] would also match the [param] pattern
-    relativePath = relativePath.replace(/\/\[\.\.\.(.*?)\]/g, "/{$1}");
-
-    // Convert Next.js dynamic route syntax to OpenAPI parameter syntax
-    relativePath = relativePath.replace(/\/\[([^\]]+)\]/g, "/{$1}");
-
-    return relativePath || "/";
   }
 
   private getSortedPaths(paths: Record<string, any>): Record<string, any> {
